#3845 comment: expanded tests
