+++ tests/unit_tests/test_interview_tasks.py
@@ -19,17 +19,16 @@ import local.dev_config  # sets env variables TEST_ON_AWS and AWS_TEST_API
 import local.secrets  # sets env variables THISCOVERY_AFS25_PROFILE and THISCOVERY_AMP205_PROFILE
 import copy
 import json
-import unittest
+import thiscovery_dev_tools.testing_tools as test_utils
+import thiscovery_lib.utilities as utils
 from http import HTTPStatus
 from pprint import pprint
+from thiscovery_lib.dynamodb_utilities import Dynamodb
 
-import thiscovery_lib.utilities as utils
 import src.endpoints as ep
-import thiscovery_dev_tools.testing_tools as test_utils
-from tests.test_data import QUALTRICS_TEST_OBJECTS, TEST_RESPONSE_DICT, ARBITRARY_UUID
-from interview_tasks import UserInterviewTask
-
+import src.common.constants as const
 import tests.test_data as td
+from interview_tasks import InterviewTask, UserInterviewTask
 
 
 class TestUserInterviewTask(test_utils.BaseTestCase):
@@ -70,3 +69,94 @@ class TestUserInterviewTask(test_utils.BaseTestCase):
             'modified',
         ]
         self.assertCountEqual(interview_task_keys, list(uit.interview_task.keys()))
+
+
+class TestUserInterviewTaskEndpoint(test_utils.BaseTestCase):
+    entity_base_url = 'v1/user-interview-tasks'
+
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        cls.ddb_client = Dynamodb(stack_name=const.STACK_NAME)
+        cls.ddb_client.delete_all(
+            table_name=const.TASK_RESPONSES_TABLE['name'],
+            key_name=const.TASK_RESPONSES_TABLE['partition_key'],
+            sort_key_name=const.TASK_RESPONSES_TABLE['sort_key'],
+        )
+        uit = UserInterviewTask(**td.TEST_USER_INTERVIEW_TASK)
+        uit.get_interview_task()
+        uit.ddb_dump()
+
+    def test_get_user_interview_task_ok(self):
+        path_parameters = {
+            'id': td.TEST_USER_INTERVIEW_TASK['response_id']
+        }
+        expected_status = HTTPStatus.OK
+
+        result = test_utils.test_get(ep.get_user_interview_task_api, self.entity_base_url, path_parameters=path_parameters)
+        result_status = result['statusCode']
+        result_body = json.loads(result['body'])
+        result_it = result_body.pop('interview_task')
+        del result_it['modified']
+        self.assertEqual(expected_status, result_status)
+        self.assertEqual(td.TEST_USER_INTERVIEW_TASK, result_body)
+        self.assertEqual(td.TEST_INTERVIEW_TASK, result_it)
+
+    def test_get_user_interview_task_does_not_exist(self):
+        path_parameters = {
+            'id': td.TEST_USER_INTERVIEW_TASK['response_id'].replace('2', '3')
+        }
+        expected_status = HTTPStatus.NOT_FOUND
+        result = test_utils.test_get(ep.get_user_interview_task_api, self.entity_base_url, path_parameters=path_parameters)
+        result_status = result['statusCode']
+        self.assertEqual(expected_status, result_status)
+
+    def test_put_user_interview_task_ok(self):
+        test_event = copy.deepcopy(td.TEST_USER_INTERVIEW_TASK_EB_EVENT)
+        result = ep.put_user_interview_task(test_event, None)
+        self.assertEqual(HTTPStatus.OK, result['statusCode'])
+
+    def test_put_user_interview_task_missing_task_id(self):
+        test_event = copy.deepcopy(td.TEST_USER_INTERVIEW_TASK_EB_EVENT)
+        del test_event['detail']['interview_task_id']
+        with self.assertRaises(utils.DetailedValueError):
+            ep.put_user_interview_task(test_event, None)
+
+
+class TestInterviewTaskEndpoint(test_utils.BaseTestCase):
+    entity_base_url = 'v1/interview-tasks'
+
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        cls.ddb_client = Dynamodb(stack_name=const.STACK_NAME)
+        cls.ddb_client.delete_all(
+            table_name=const.INTERVIEW_TASKS_TABLE['name'],
+            key_name=const.INTERVIEW_TASKS_TABLE['partition_key'],
+            sort_key_name=const.INTERVIEW_TASKS_TABLE['sort_key'],
+        )
+        test_it = InterviewTask(**td.TEST_INTERVIEW_TASK)
+        test_it.ddb_dump()
+
+    def test_get_interview_task_ok(self):
+        path_parameters = {
+            'id': td.TEST_INTERVIEW_TASK['interview_task_id']
+        }
+        expected_status = HTTPStatus.OK
+        result = test_utils.test_get(ep.get_interview_task_api, self.entity_base_url, path_parameters=path_parameters)
+        result_status = result['statusCode']
+        result_body = json.loads(result['body'])
+        del result_body['modified']
+        self.assertEqual(expected_status, result_status)
+        self.assertEqual(td.TEST_INTERVIEW_TASK, result_body)
+
+    def test_get_interview_task_not_found(self):
+        path_parameters = {
+            'id': td.ARBITRARY_UUID
+        }
+        expected_status = HTTPStatus.NOT_FOUND
+        result = test_utils.test_get(ep.get_interview_task_api, self.entity_base_url, path_parameters=path_parameters)
+        result_status = result['statusCode']
+        self.assertEqual(expected_status, result_status)
+
+
